backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.0
python-multipart==0.0.6
SQLAlchemy==2.0.19
Pillow==10.0.0
numpy==1.25.1
ftfy==6.1.1
regex==2023.6.3
torch==2.0.1
transformers==4.30.2
torch
pillow
numpy
tqdm
exifread
geopy
git+https://github.com/openai/CLIP.git
argparse
pathlib